TeamsAPI.getApp(for:with:)
